scst: Fix recently introduced Coverity warnings
Add error handling for recently introduced SHA256 and SHA3-256 APIs.

+++ qla2x00t/qla2x00-target/qla2x00t.c
@@ -5841,7 +5841,7 @@ out_free_fcport:
 	kfree(fcport);
 
 out:
-	TRACE_EXIT_HRES((unsigned long)sess);
+	TRACE_EXIT_HRES(sess);
 	return sess;
 }
